Refactor app.py to create an application factory and improve environment variable handling; update RepoDTO to allow optional description; modify getRepos method signature; remove obsolete test files and add new test cases for GitHub service and Flask app.
